move render start to helper
