+++ tasty-application/src/utils/helpers.js
@@ -1,5 +1,24 @@
-export const navigateToRecipe = (navigate, id) => {
+import React from 'react'
+
+// Helper function to render star ratings
+  const renderStars = (score) => {
+    const maxStars = 5
+    const filledStars = Math.round(score * maxStars) // Convert score (0-1) to stars (0-5)
+    return (
+      <div className="recommend-card-stars d-flex justify-content-center mb-0">
+        {[...Array(maxStars)].map((_, index) => (
+          <span key={index} className={index < filledStars ? 'star filled' : 'star'}>
+            ★
+          </span>
+        ))}
+      </div>
+    )
+  }
+
+const navigateToRecipe = (navigate, id) => {
     if (id) {
       navigate(`/recipe/${id}`)
     }
   }
+
+export { renderStars, navigateToRecipe }
